requirements-test.txt
pytest>=8.0
pytest-cov
python-frontmatter>=1.1
pyyaml>=6.0
anthropic>=0.40
